requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.2
python-dotenv==1.0.1
